requirements.txt
fastapi
uvicorn
python-dotenv
requests
pyjwt
cryptography